+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Social Share Change Log
 * Enh: Added `driverProperties` property to widget for adding of
 custom properties to drivers
 * Enh: Creates unit tests for `Tumblr` driver
+* Fix: Using `Html::tag()` instead of `Html::a()` in widget
 
 1.3.0 October 02, 2017
 ----------------------