+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Social Share Change Log
 custom properties to drivers
 * Enh: Creates unit tests for `Tumblr`, `Twitter`, `LinkedIn`, `Telegram`,
  `Facebook`, `Gmail` drivers
+* Enh: Using minified version of CSS file for social icons on production
 * Fix: Using `Html::tag()` instead of `Html::a()` in widget
 * Fix: Call parent implementation of method `init()` in widget
 * Fix: Encode `body` param in Gmail driver